create split modules relates #37
